requirements.txt
ultralytics>=8.2.0
opencv-python>=4.9.0
numpy>=1.24.0
pillow>=10.0.0
torch>=2.1.0
 torchvision>=0.16.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
matplotlib>=3.8.0
seaborn>=0.13.0
rich>=13.7.0
tqdm>=4.66.0
omegaconf>=2.3.0
scikit-learn>=1.3.0
